+++ jni/src/main/java/org/derecalliance/derec/crypto/DerecCryptoInterface.java
@@ -1,4 +1,4 @@
-package src;
+package org.derecalliance.derec.crypto;
 
 import java.util.List;
